requirements.txt
pandas
numpy
sklearn
scipy
pyaml